Major edit: Restructure bonding components
