+++ src/components/Chart/index.js
@@ -1,1 +1,1 @@
-export * from './D3Chart'
+
